Compared and changed customstylesheet.css
Compared and changed the customstylesheet.css to the customstylesheet_acceptatie.css to make sure they are not different to each other anymore.

Also fixed a typo where there was no closing bracket on line 477 in customstylesheet.css. The same problem was on _acceptatie but on a different line.

It is important that we follow the flow:
1. Acceptance fetch origin.
2. Make changes locally on acceptance.
3. Commit to acc-jachensen.nl
4. Make sure everything works as expected.
5. Fetch from origin
6. Copy changes to production css locally.
7. commit changes to github
8. test changes again on jachensen.nl